kernel: Update kposix lib

+++ sys/kernel/include/posix/string.h
@@ -6,6 +6,6 @@
 #include <sys/types.h>
 
 void *memcpy(void *s1, const void *s2, size_t n);
-void *memset(void *ptr, int val, size_t length);
+void *memset(void *ptr, uint8_t val, size_t length);
 
 #endif
