+++ sys/kernel/posix/memset.c
@@ -2,7 +2,7 @@
 
 #include <string.h>
 
-void *memset(void *ptr, int val, size_t length)
+void *memset(void *ptr, uint8_t val, size_t length)
 {
   for (size_t i = 0; i < length; ++i)
   {
